#1441 extracting contains-file logic
